+++ Anchor.toml
@@ -1,5 +1,4 @@
 [toolchain]
-channel="nightly"
 
 [features]
 resolution = true
@@ -43,6 +42,7 @@ claim-quest = "7nsn4z8U1nVCVHud9CLmYLy5ZHK2bSMge6u7YgmssdaA"
 claim_loot = "4CjxHvNUpoCYomULBFTvmkTQPaNd9QDHPhZQ6eB9bZEf"
 claim_time = "4XXA1mX5aN4Fd62FBgNxCU7FzKDYS3KSxFX3RdJYoWPj"
 deity-bot = "62f9zAUjCN5VFqWF43qSUrW6CvivqhsEjDvCHwQ1SjgR"
+ecs-bundle = "Cjca6tWWGx77ki6rRinErcoZJEvJHNxfPaut8DoKJHQ5"
 exchange = "Csna3V2jUMdQEQKUCxLsQEnYThAGPSWcPCxW9vea1S8d"
 hero = "GBzY8ujNDb1FNkJUXUUjKV5uZPqzi6AoKsPjsqFEHCeh"
 location_allocator = "J7q3dEg2KauPKkMamH9Q5FHhCoFYsSq9ramdutMpPTDc"
@@ -70,11 +70,11 @@ worker = "BExuAEwcKxKeqHSN8C1WetUAd6Tm71cZEiP8EBSrH55T"
 url = "https://api.apr.dev"
 
 [provider]
-cluster = "Localnet"
+cluster = "localnet"
 wallet = "~/.config/solana/id.json"
 
 [workspace]
-members = ["programs-ecs/components/*", "programs-ecs/systems/*", "programs/*"]
+members = ["programs-ecs/components/*", "programs-ecs/systems/*", "programs-ecs/bundles/*", "programs/*"]
 
 [scripts]
 test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.test.ts"
