+++ programs-ecs/bundles/ecs-bundle/src/lib.rs
@@ -1,0 +1,29 @@
+use bolt_lang::*;
+
+pub mod settlement;
+
+declare_id!("Cjca6tWWGx77ki6rRinErcoZJEvJHNxfPaut8DoKJHQ5");
+
+#[bundle]
+pub mod ecs_bundle {
+	use crate::settlement::{Building, EnvironmentState, ResourceBalance};
+
+	#[component(delegate)]
+	pub struct Settlement {
+		#[max_len(20, 6)]
+		pub buildings: Vec<Building>,
+		pub owner: Pubkey,
+
+		pub environment: EnvironmentState,
+		pub treasury: ResourceBalance,
+
+		pub faith: u8,
+		pub time_units: u8,
+		pub last_time_claim: i64,
+		pub research: u32,
+
+		#[max_len(30, 1)]
+		pub worker_assignment: Vec<i8>, //index is worker unit index, value is building index from /buildings/ array; singed: use -1 for free slot
+		pub quest_claim_status: u64,
+	}
+}
