+++ programs-ecs/bundles/ecs-bundle/src/settlement/mod.rs
@@ -3,6 +3,8 @@ pub mod config;
 use bolt_lang::*;
 use config::{get_extraction_cap, BuildingType};
 
+pub use crate::ecs_bundle::{Settlement, SettlementInit};
+
 declare_id!("5bKBE1HgusXC5jVVjpk4CvxUM8UGnVPQyvGt7cB6Jk7W");
 
 #[component_deserialize]
@@ -39,25 +41,6 @@ pub struct Building {
     pub level: u8,
     pub turns_to_build: u8,
     pub extraction: u16, //mines only
-}
-
-#[component(delegate)]
-pub struct Settlement {
-    #[max_len(20, 6)]
-    pub buildings: Vec<Building>,
-    pub owner: Pubkey,
-
-    pub environment: EnvironmentState,
-    pub treasury: ResourceBalance,
-
-    pub faith: u8,
-    pub time_units: u8,
-    pub last_time_claim: i64,
-    pub research: u32,
-
-    #[max_len(30, 1)]
-    pub worker_assignment: Vec<i8>, //index is worker unit index, value is building index from /buildings/ array; singed: use -1 for free slot
-    pub quest_claim_status: u64,
 }
 
 impl Default for Settlement {
