+++ programs-ecs/systems/change-backpack/Cargo.toml
@@ -20,5 +20,5 @@ idl-build = ["anchor-lang/idl-build"]
 bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
-settlement = { version = "0.1.10", path = "../../components/settlement", features = ["cpi"] }
+ecs-bundle.workspace = true
 hero = { version = "0.1.10", path = "../../components/hero", features = ["cpi"] }
