+++ programs-ecs/systems/claim-quest/src/lib.rs
@@ -5,11 +5,11 @@ declare_id!("7nsn4z8U1nVCVHud9CLmYLy5ZHK2bSMge6u7YgmssdaA");
 
 #[system]
 pub mod claim_quest {
-    use settlement::config::get_quest_progress;
+    use ecs_bundle::settlement::config::get_quest_progress;
 
-    use settlement::config::Resource;
-    use settlement::config::QUESTS_CONFIG;
-    use settlement::Settlement;
+    use ecs_bundle::settlement::config::Resource;
+    use ecs_bundle::settlement::config::QUESTS_CONFIG;
+    use ecs_bundle::settlement::Settlement;
 
     use crate::errors::QuestClaimError;
