+++ programs-ecs/systems/exchange/Cargo.toml
@@ -22,4 +22,4 @@ anchor-lang.workspace = true
 serde.workspace = true
 anchor-spl.workspace = true
 token-minter = { version = "0.1.0", path = "../../../programs/token-minter", features = ["cpi"] }
-settlement = { version = "0.1.10", path = "../../components/settlement", features = ["cpi"] }
+ecs-bundle.workspace = true
