+++ programs-ecs/systems/research/Cargo.toml
@@ -20,4 +20,4 @@ idl-build = ["anchor-lang/idl-build"]
 bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
-settlement = { version = "0.1.10", path = "../../components/settlement", features = ["cpi"] }
+ecs-bundle.workspace = true
