+++ programs-ecs/systems/wait/Cargo.toml
@@ -21,4 +21,4 @@ bolt-lang.workspace = true
 anchor-lang.workspace = true
 anchor-spl.workspace = true
 serde.workspace = true
-settlement = { version = "0.1.10", path = "../../components/settlement", features = ["cpi"] }
+ecs-bundle.workspace = true
