+++ tests/exploration.test.ts
@@ -6,6 +6,7 @@ import { LocationAllocatorWrapper } from "./wrappers/location_allocator.wrapper"
 import { TokenWrapper } from "./wrappers/token.wrapper";
 import { HeroWrapper } from "./wrappers/hero.wrapper";
 import { LootDistributionWrapper } from "./wrappers/loot_distribution.wrapper";
+import { Component } from "../../bolt/clients/typescript/lib";
 
 
 
@@ -33,7 +34,7 @@ describe("Exploration and multiplayer tests", async () => {
         await settlement.init(await world.getWorldPda(), allocatorState.currentX, allocatorState.currentY)
         const state = await player.assignSettlement(
             settlement.entityPda,
-            settlement.settlementComponent.programId,
+            new Component(settlement.bundle.programId, "settlement"),
             locationAllocator.entityPda,
             locationAllocator.locationAllocatorComponent.programId,
         );
@@ -123,7 +124,7 @@ describe("Exploration and multiplayer tests", async () => {
         const heroBefore = await hero.state();
 
         await hero.changeBackpack(settlement.entityPda,
-            settlement.settlementComponent.programId, { water: 0, food: 0, wood: 1, stone: 0 });
+            new Component(settlement.bundle.programId, "settlement"), { water: 0, food: 0, wood: 1, stone: 0 });
 
 
         const settlementAfter = await settlement.state();
