+++ tests/wrappers/token.wrapper.ts
@@ -3,7 +3,7 @@ import { AccountMeta, PublicKey } from "@solana/web3.js";
 import {
   FindComponentPda,
   InitializeNewWorld,
-} from "@magicblock-labs/bolt-sdk"
+} from "../../../bolt/clients/typescript/lib"
 
 
 import { TokenMinter } from "../../target/types/token_minter";
